+++ tests/test_6_lhc.py
@@ -1,7 +1,7 @@
 import numpy as np
 
 import nafflib
-from nafflib.ndim import linear_combinations
+from nafflib.indexing import linear_combinations
 
 # =================================
 num_turns = int(3e4)
